Simplify RetryManager and move to E2E package
