--- src/main/java/teammates/common/util/retry/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Contains infrastructure and helpers needed for retrying arbitrary code.
- */
-package teammates.common.util.retry;
